+++ features/support/hooks.rb
@@ -1,0 +1,3 @@
+Before('@no-travis') do
+  skip_this_scenario if ENV['CI']
+end